Struct leetcode::Solution

pub struct Solution;

Implementations

impl Solution

pub fn circle_game(
toys: &Vec<(usize, usize, usize)>,
circles: &Vec<(usize, usize)>,
radius: usize
) -> usize

LCP 42 - Circle Game (772 ms 3.7 MB)

toy (x, y, r), cycle (x, y). Simply traverse the toys smaller than circles for each circle and filter toys in circles then return the count after filter. A toy in a circle if their distance of centres isn’t greater than their radius subtraction.

pub fn find_restaurant(list1: Vec<String>, list2: Vec<String>) -> Vec<String>

599 - Minimum Index Sum of Two Lists (8 ms 2.3 MB)

Suppose Andy and Doris want to choose a restaurant for dinner, and they both have a list of favorite restaurants represented by strings.

You need to help them find out their common interest with the least list index sum. If there is a choice tie between answers, output all of them with no order requirement. You could assume there always exists an answer.

Firstly, we entry the list2 into a hash map. Then we traverse the list1 and find the same keys from list2. We first guess the min index sum is usize::MAX, then if a new index sum less than old one, we clear the result and then push the key into it, or a index sum is same as the min we just push key into it but not clear the result.

pub fn best_rotation(nums: &Vec<usize>) -> usize

798 - Smallest Rotation with Highest Score (4 ms 4 MB)

You are given anarray nums. You can rotate it by a non-negative integer k so that the array becomes [nums[k], nums[k + 1], … nums[nums.length - 1], nums[0], nums[1], …, nums[k-1]]. Afterward, any entries that are less than or equal to their index are worth one point.

For example, if we have nums = [2, 4, 1, 3, 0], and we rotate by k = 2, it becomes [1, 3, 0, 2, 4]. This is worth 3 points because 1 > 0 [no points], 3 > 1 [no points], 0 <= 2 [one point], 2 <= 3 [one point], 4 <= 4 [one point].

Return the rotation index k that corresponds to the highest score we can achieve if we rotated nums by it. If there are multiple answers, return the smallest such index k.

Constraints:

1 <= nums.length <= 10^5

0 <= nums[i] < nums.length

Firstly, we map the diffs of each num, and calculate the score of k = 0. For diff > 0, we map the nums to it’s last k it could get a point. Then we traverse k from 1 to the end. There are many changes before caculate the next score. First step is get current diff (which will move to the end). If it has got a score, then minus a point previous (it won’t lose a point after moving to the end), others add a point to it’s first k that got a point. Note that a negative diff (the diff after apply the k is negative) will always get a point after a rotation. Then we minus the score will lost after a rotation. After changes, k += 1 (a rotation has finished). If current score greater than max score, then update the result.

pub fn freq_alphabets(str: &str) -> String

1309 - Decrypt String from Alphabet to Integer Mapping (0 ms 2 MB)

You are given a string str formed by digits and '#'. We want to map str to English lowercase characters as follows:

Characters ('a' to 'i') are represented by ('1' to '9') respectively.

Characters ('j' to 'z') are represented by ('10#' to '26#') respectively.

Firstly, we create a queue which map to the high and the low. Then we traverse the str. When a byte is '#', we pop both the high and the low as a char, others we push the byte into the queue and pop the high as a char. If the queue isn’t empty after traversal, pop the remaining bytes. After all steps, we build a dist string successfully.

pub fn plates_between_candles(
str: &str,
queries: &Vec<(usize, usize)>
) -> Vec<usize>

2055 - Plates Between Candles (48 ms 9.6 MB)

There is a long table with a line of plates and candles arranged on top of it. You are given a 0-indexed string str consisting of characters '*' and '|' only, where a '*' represents a plate and a '|' represents a candle.

You are also given a 0-indexed 2D integer array queries where queries[i] = (left, right) denotes the substring str[left..=right]. For each query, you need to find the number of plates between candles that are in the substring. A plate is considered between candles if there is at least one candle to its left and at least one candle to its right in the substring.

For solving this problem, we can simply traversestr for each query, but it’s sloooooowly (O(m * n)). We have to find an alternative. Firstly, we can get the left plates count of each index of str. Then we get the closest candle index both the left side and right side of each index of str. Finally we minus the left plates count of the closest left candle index of query[right] and the left plates count of the closest right candle index of query[left], but we need to verify whether both the closest left candle index of query[right] and the closest right candle index of query[left] are valid. (which actually O(3 * m + n))